Made NPC Drops Dynamic
Former-commit-id: 6ca65242fc6d7194a19eff9dacf68b4d08c21431

+++ Intersect Editor/Forms/Editors/frmResource.cs
@@ -193,7 +193,7 @@ namespace Intersect.Editor.Classes
             cmbItem.SelectedIndex =
                 Database.GameObjectListIndex(GameObjectType.Item, mEditorItem.Drops[index].ItemNum) + 1;
             nudDropAmount.Value = mEditorItem.Drops[index].Amount;
-            nudDropChance.Value = mEditorItem.Drops[index].Chance;
+            nudDropChance.Value = (decimal)mEditorItem.Drops[index].Chance;
         }
 
         private void nudSpawnDuration_ValueChanged(object sender, EventArgs e)
